(refactor) renamed the useHWR.ts to searchHealthCareWork.ts
